refactor: decompose Canvas.svelte markup and swipe handlers (#1610)
Extract the rack-rendering markup into RackCanvasView and the mobile
swipe-to-switch-rack state machine into a canvas-swipe controller, leaving
Canvas.svelte as a viewport shell. Coordinate math and panzoom lifecycle were
already extracted in #2111.

- Canvas.svelte: 988 -> 471 LOC (<500 target met)
- RackCanvasView.svelte (new): rack markup + device/placement/select handlers
- canvas-swipe.svelte.ts (new): swipe gesture state machine
- canvas-swipe.test.ts (new): 8 controller tests for gating and switching

No behaviour change. Imports stay at 18: the <12 sub-target is not reachable
without a barrel convention the codebase does not use (see issue note).
